refactor: improve layout part 1
