+++ src/components/UserInfo.tsx
@@ -1,8 +1,11 @@
+import { Box } from '@mui/material';
 import React from 'react';
 
 interface UserInfoProps {
   currentProfileKey: ProfileDetailRawDataUnionType;
   currentProfileValue?: string | number;
+  profileDetail: ProfileDetailRawDataUnionType[];
+  profiledetailIndex: number;
 }
 
 export enum ProfileDetailRawData {
@@ -16,16 +19,21 @@ export enum ProfileDetailRawData {
 
 export type ProfileDetailRawDataUnionType = `${ProfileDetailRawData}`;
 
-export default function UserInfo({ currentProfileKey, currentProfileValue }: UserInfoProps) {
+export default function UserInfo({ currentProfileKey, currentProfileValue, profileDetail, profiledetailIndex }: UserInfoProps) {
   const camelCaseKey = `${currentProfileKey[0].toUpperCase()}${currentProfileKey.substring(1, currentProfileKey.length)}`;
 
   return (
     <>
       {currentProfileKey && currentProfileValue ? (
-        <div className="mb-2.5">
-          <span className="mr-4 font-medium	text-[#555]">{camelCaseKey}</span>
-          <span className="font-light">{currentProfileValue}</span>
-        </div>
+        <Box
+          display="flex"
+          flexDirection="row"
+          mb={profiledetailIndex === profileDetail.length - 1 ? 0 : 1}
+        >
+          <h3 className="mr-4 font-medium	text-[#555] flex-[4]">{camelCaseKey}</h3>
+
+          <p className="font-light flex-[8]">{currentProfileValue}</p>
+        </Box>
       ) : null}
     </>
   );
